Starting files of the work in the v2.0 node
I started drawing some new features in the schematic for the v2.0 node. I really wanna test a 18650 + MPTT charger on board, also testing the possibility of a nice RTC on the other side of the board for nodes that maybe need them. This is going to be a major rework of the Node board.